refactor(payment_methods): move `PaymentsMandateReference` from `router` crate to `diesel_models` crate
